separate out components, add new algorithms
update dependencies
