+++ pyproject.toml
@@ -13,7 +13,7 @@ dependencies = [
     "numpy",
     "matplotlib",
 ]
-python_version == "3.12.*"
+requires-python = ">=3.12, <3.13"
 
 [project.optional-dependencies]
 dev = [
