+++ src/components/figure.py
@@ -76,37 +76,4 @@ def draw_plot(draw_plot, plotoptions, relayoutData, browser_data, graphData, hol
     return fig
 
 
-@callback(
-    Output("draw-plot", "data", allow_duplicate=True),
-    Output("click-data", "children", allow_duplicate=True),
-    Output("ui", "children", allow_duplicate=True),
-    Output("browser-data", "data", allow_duplicate=True),
-    Output("graph-data", "data", allow_duplicate=True),
-    Output("holes-data", "data", allow_duplicate=True),
-    Input("reset", "n_clicks"),
-    State("xmax", "value"),
-    State("ymax", "value"),
-    State("zmax", "value"),
-    prevent_initial_call=True,
-)
-def reset_grid(n_clicks, xslider, yslider, zslider):
-    """
-    Reset the grid.
-    """
-    s = BrowserState()
-    s.xmax = int(xslider)
-    s.ymax = int(yslider)
-    s.zmax = int(zslider)
-    s.shape = [s.xmax, s.ymax, s.zmax]
-    s.removed_nodes = np.zeros(s.xmax * s.ymax * s.zmax, dtype=bool)
-    G = Grid(s.shape)
-    D = Holes(s.shape)
-    # Make sure the view/angle stays the same when updating the figure
-    return (
-        1,
-        s.log,
-        "Created grid of shape {}".format(s.shape),
-        jsonpickle.encode(s),
-        G.encode(),
-        D.encode(),
-    )
+
